requirements.txt
# Black Ops Framework Requirements
# Core Dependencies
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
colorama>=0.4.6
cryptography>=38.0.0
PyYAML>=6.0

# Network & Security
scapy>=2.5.0
dnspython>=2.3.0
paramiko>=3.0.0
netifaces>=0.11.0
python-nmap>=0.7.1

# Data Processing
pandas>=1.5.0
numpy>=1.23.0
openpyxl>=3.0.0
python-docx>=0.8.11
reportlab>=3.6.0

# Web & API
selenium>=4.8.0
webdriver-manager>=3.8.0
tweepy>=4.12.0
instagram-private-api>=1.6.0

# Utilities
progress>=1.6
tabulate>=0.9.0
python-magic>=0.4.27
psutil>=5.9.0
pyfiglet>=0.8.post1

# Testing
pytest>=7.2.0
pytest-cov>=4.0.0

# Optional (for specific tools)
pycryptodome>=3.17
hashid>=3.1.4
whois>=0.9.0
geoip2>=4.6.0
exif>=1.6.0